Item Serpong

+++ Core/DomainModel/StockAndMaster/WarehouseItem.cs
@@ -12,7 +12,7 @@ namespace Core.DomainModel
         public int ItemId { get; set; }
         //public int CustomerItemId { get; set; }
 
-        public int Quantity { get; set; } // Company's Item Quantity (excluding customer's)
+        public decimal Quantity { get; set; } // Company's Item Quantity (excluding customer's)
         public int PendingDelivery { get; set; }
         public int PendingReceival { get; set; }
         public int CustomerQuantity { get; set; } // Customer's Item Quantity
