+++ Core/Interface/Service/StockAndMaster/IItemService.cs
@@ -35,18 +35,16 @@ namespace Core.Interface.Service
                                     IBlanketService _blanketService, IPurchaseOrderDetailService _purchaseOrderDetailService,
                                     IStockAdjustmentDetailService _stockAdjustmentDetailService, ISalesOrderDetailService _salesOrderDetailService,
                                     IPriceMutationService _priceMutationService, IBlanketOrderDetailService _blanketOrderDetailService);
-        Item AdjustCustomerQuantity(Item item, int quantity);
+        Item AdjustCustomerQuantity(Item item, decimal quantity);
         //Item AdjustCustomerVirtual(Item item, int quantity);
-        Item AdjustQuantity(Item item, int quantity);
+        Item AdjustQuantity(Item item, decimal quantity);
         Item AdjustPendingReceival(Item item, int quantity);
         Item AdjustPendingDelivery(Item item, int quantity);
         Item AdjustVirtual(Item item, int quantity);
-        //Item OnTrial(Item item, int quantity);
-        //Item ReturnTrial(Item item, int quantity);
-        decimal CalculateAvgPrice(Item item, int addedQuantity, decimal addedAvgPrice);
-        decimal CalculateAndUpdateAvgPrice(Item item, int addedQuantity, decimal addedAvgPrice);
-        decimal CalculateCustomerAvgPrice(Item item, int addedQuantity, decimal addedAvgPrice);
-        decimal CalculateAndUpdateCustomerAvgPrice(Item item, int addedQuantity, decimal addedAvgPrice);
+        decimal CalculateAvgPrice(Item item, decimal addedQuantity, decimal addedAvgPrice);
+        decimal CalculateAndUpdateAvgPrice(Item item, decimal addedQuantity, decimal addedAvgPrice);
+        decimal CalculateCustomerAvgPrice(Item item, decimal addedQuantity, decimal addedAvgPrice);
+        decimal CalculateAndUpdateCustomerAvgPrice(Item item, decimal addedQuantity, decimal addedAvgPrice);
         bool DeleteObject(int Id);
         bool IsSkuDuplicated(Item item);
     }
