+++ Validation/Validation/StockAndMaster/StockAdjustmentDetailValidator.cs
@@ -52,10 +52,12 @@ namespace Validation.Validation
 
         public StockAdjustmentDetail VNonZeroNorNegativePrice(StockAdjustmentDetail stockAdjustmentDetail)
         {
+            /* TODO: Skip Price = 0 Untuk data awal
             if (stockAdjustmentDetail.Price <= 0)
             {
                 stockAdjustmentDetail.Errors.Add("Price", "Harus lebih besar dari 0");
             }
+             */
             return stockAdjustmentDetail;
         }
 
@@ -104,7 +106,7 @@ namespace Validation.Validation
         public StockAdjustmentDetail VNonNegativeStockQuantity(StockAdjustmentDetail stockAdjustmentDetail, IStockAdjustmentService _stockAdjustmentService,
                                                                IItemService _itemService, IBlanketService _blanketService, IWarehouseItemService _warehouseItemService, bool ToConfirm)
         {
-            int stockAdjustmentDetailQuantity = ToConfirm ? stockAdjustmentDetail.Quantity : ((-1) * stockAdjustmentDetail.Quantity);
+            decimal stockAdjustmentDetailQuantity = ToConfirm ? stockAdjustmentDetail.Quantity : ((-1) * stockAdjustmentDetail.Quantity);
             //decimal stockAdjustmentDetailPrice = ToConfirm ? stockAdjustmentDetail.Price : ((-1) * stockAdjustmentDetail.Price);
             Item item = _itemService.GetObjectById(stockAdjustmentDetail.ItemId);
             StockAdjustment stockAdjustment = _stockAdjustmentService.GetObjectById(stockAdjustmentDetail.StockAdjustmentId);
